refactor(cli): flatten mcp command module
Move the MCP command slice to the neutral command root and update its internal export and approved test fixtures without changing MCP behavior.
